uart-hdl: uart_rx: Use report to run test benches
UART rx test bench now instanitates a fifo to write data and errors
from the receiver. The test bench now automatically reads from this
fifo to see if data is correct and errors are either generated or
not generated correctly. Report statements are used to report any
failures in the test case otherwise success is printed.

Signed-off-by: Gregory Williams <gregwills85@gmail.com>