uart-hdl: Add tcl script to generate XSA with vivado
Adds tcl script to generate full XSA for the uart device. Block design
is created with ZYNQ7 processing system and connects the M_AXI_0 port to
the S_AXI port of uart device. Also connects external IO to PMOD port on
the zybo board and connects uart_intr to the PL-PS interrupt input.

Signed-off-by: Gregory Williams <gregwills85@gmail.com>